third_party/blink/web_tests/inspector-protocol/css/css-get-platform-fonts-display-locked-expected.txt

Test css.getPlatformFontsForNode method with display locking.

Running test: testParent

Running test: testChild
